--- a/src/js/ia/IADevPipeline.js
+++ b/src/js/ia/IADevPipeline.js
@@ -28,6 +28,12 @@
             var url = window.location.pathname.replace(/\/$/, '') + "/json";
             var username = $(".user-name").text();
 
+	    // 100% width
+            $(".site-main > .content-wrap").first().removeClass("content-wrap").addClass("wrap-pipeline");
+	    $(".breadcrumb-nav").remove();
+	    $(".site-main").addClass("developer-main");
+
+
             $.getJSON(url, function(data) { 
                 // console.log(window.location.pathname);
                 // Check user permissions and add to the data
@@ -52,11 +58,6 @@
                     $("#select-milestone").addClass("hide");
                     $("#select-type").removeClass("hide");
                 }
-
-                // 100% width
-                $(".site-main > .content-wrap").first().removeClass("content-wrap").addClass("wrap-pipeline");
-		$(".breadcrumb-nav").remove();
-		$(".site-main").addClass("developer-main");
 
                 var parameters = window.location.search.replace("?", "");
                 parameters = $.trim(parameters.replace(/\/$/, ''));