--- a/src/scss/ia/_deprecated.scss
+++ b/src/scss/ia/_deprecated.scss
@@ -1,3 +1,11 @@
+#deprecated {
+    @include center-content();
+}
+
+.deprecated-wrap {
+    @include wrapper();
+}
+
 #deprecated, #ghosted {
     .dev_pipeline-column {
         width: 15.5%;