--- a/src/scss/ia/_dev-nav.scss
+++ b/src/scss/ia/_dev-nav.scss
@@ -1,0 +1,30 @@
+// This is the "blue bar" navigation that
+// is available across all the developer pages.
+.developer-nav {
+    background: $blue-light;
+    padding: 1em 1em;
+    height: 150px;
+    color: white;
+
+    // Adjust z-index to be above the sidebar.
+    position: relative;
+    z-index: $pipeline-blue-bar;
+    
+    h1 {
+	color: currentColor;
+    }
+
+    .pages {
+	text-align: center;
+
+	& > a {
+	    padding: 0.5em;
+	    color: currentColor;
+	    opacity: .45;
+	}
+
+	& > a.active-link {
+	    opacity: 1;
+	}
+    }
+}