--- a/src/scss/ia/_dev-nav.scss
+++ b/src/scss/ia/_dev-nav.scss
@@ -2,8 +2,8 @@
 // is available across all the developer pages.
 .developer-nav {
     background: $blue-light;
-    padding: 1em 1em;
-    height: 150px;
+    padding: 2em 1em;
+    height: 130px;
     color: white;
 
     // Adjust z-index to be above the sidebar.
@@ -12,13 +12,14 @@
     
     h1 {
 	color: currentColor;
+	margin-bottom: 0.5em;
     }
 
     .pages {
 	text-align: center;
 
 	& > a {
-	    padding: 0.5em;
+	    padding: 1em;
 	    color: currentColor;
 	    opacity: .45;
 	}