--- a/src/scss/ia/_ia.scss
+++ b/src/scss/ia/_ia.scss
@@ -1,3 +1,8 @@
+// This removes the padding from the .site-main class.
+.developer-main {
+    padding: 0;
+}
+
 /*te-new-ia.button experimental and/or prototype */
 
 .half, .fifty {
@@ -1172,18 +1177,6 @@ input.not_saved, select.not_saved, .button.not_saved {
 .filter-issues {
     width: 23%;
     margin-left: 3em;
-}
-
-#pipeline_toggle-view {
-    margin-bottom: 1em;
-}
-
-#pipeline_toggle-view .pager-wrap {
-    height: 4em;
-}
-
-#pipeline_toggle-view .pipeline__toggle-view__button {
-    box-shadow: none;
 }
 
 #pipeline_toggle-dev {