--- a/src/scss/ia/_mixins.scss
+++ b/src/scss/ia/_mixins.scss
@@ -597,3 +597,12 @@ $remsize: 16 * (parseInt($html-font-size) / 100);
 		content: "\4F";
 	}
 }
+
+@mixin wrapper() {
+    /* Make the fonts look snazzy. */
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+
+    font-size: 13px;
+    min-height: 100vh;
+}