--- a/src/scss/ia/_overview.scss
+++ b/src/scss/ia/_overview.scss
@@ -1,6 +1,4 @@
 @import "../../../bower_components/duckduckgo-colors/_colors";
-@import "compass/css3/border-radius";
-@import "compass/css3/text-shadow";
 
 .heavy {
     font-weight: 700;
@@ -91,14 +89,14 @@
 
         .tag-container { 
             padding: 0.15em 0.5em; 
-            @include border-radius(3px);
+            border-radius: 3px;
             text-align: center;
             max-width: 15em;
             min-width: 5em;
 
             h3 { 
                 margin-top: 0.25em; 
-                @include text-shadow(2px 2px 2px $slate);
+                text-shadow: 2px 2px 2px $slate;
             }
         }