--- a/src/scss/ia/_pipeline.scss
+++ b/src/scss/ia/_pipeline.scss
@@ -1,4 +1,4 @@
-$blue-variant: #7AB4EE;
+$blue-variant: #7AB4E3;
 $gold-variant: #F6BE68;
 
 #sort_pipeline {