--- a/src/scss/ia/_pipeline.scss
+++ b/src/scss/ia/_pipeline.scss
@@ -367,7 +367,7 @@ textarea.edit-sidebar {
 	}
 
 	&.dog-ear {
-	    @include background-image(linear-gradient(
+	    @include background(linear-gradient(
 					  -135deg, 
 					  $blue-variant,
 					  $blue-variant 2%,
@@ -407,26 +407,26 @@ textarea.edit-sidebar {
 	    }
 
             /* Change content to match the repo. */
-            &.pipeline__stamp {
+            &.pipeline__stamp- {
 
-		&-:before {
+		&:before {
 		    @include pipeline_stamp();
 		    margin-left: -3px;
 		}
 
-                &-spice:before,  &-goodies:before, &-fathead:before, &-longtail:before {
+                &spice:before,  &goodies:before, &fathead:before, &longtail:before {
 		    @include pipeline_stamp();
                 }
 
-		&-goodies:before {
+		&goodies:before {
 		    margin-left: -2px;
 		}
 
-		&-spice:before {
+		&spice:before {
 		    margin-left: -4px;
 		}
 
-		&-fathead:before, &-longtail:before {
+		&fathead:before, &longtail:before {
 		    margin-left: -5px;
 		}
             }