--- a/src/scss/ia/_pipeline.scss
+++ b/src/scss/ia/_pipeline.scss
@@ -346,7 +346,7 @@ textarea.edit-sidebar {
         list-style: none;
         padding: 0.3em 0;
         font-size: 1em;
-        background-color: black;
+        background-color: white;
         margin-bottom: 0.5em;
         border-radius: 2px;
         padding: 0.5em;