--- a/src/scss/ia/main.scss
+++ b/src/scss/ia/main.scss
@@ -21,6 +21,7 @@
 @import "../z-index";
 
 @import "dev-nav";
+@import "mixins";
 
 @import "fonts";
 @import "ia";