… bower.json
… package.json
… root/static/css/content.css
… templates/instantanswer/deprecated.tx
… templates/instantanswer/dev_pipeline.tx
… templates/instantanswer/issues.tx
… templates/instantanswer/overview.tx
… templates/scripts.tx