In preparation for the improvements that we need to make for GQ&H (https://projects.invisionapp.com/share/5T52AX88Q0), I'd like to make a small addition to grunt to improve speeds. Currently, I'd have to wait for 6 seconds for `grunt watch` to build all the CSS files, and it starts to get annoying after that.

It's a WIP since I have yet to test if everything is working as before, but preliminary results on my machine have cut down the time from 6s to 1s. Here's what had to be done for the speedup:

1. Remove unnecessary compiling of Handlebars and JS files when only the CSS changed. (cut down time from 6s to 4s)
2.  Move to libsass. (cut down from 4s to 1s)

<img width="1418" alt="screen_shot_2015-11-26_at_3_36_44_pm" src="https://cloud.githubusercontent.com/assets/81969/11430904/be114866-945b-11e5-9994-e88dcadaf9c4.png">
<img width="1418" alt="screen_shot_2015-11-26_at_3_51_02_pm" src="https://cloud.githubusercontent.com/assets/81969/11430905/be1549f2-945b-11e5-9c88-83f92b3196a7.png">
<img width="1418" alt="screen_shot_2015-11-26_at_4_11_00_pm" src="https://cloud.githubusercontent.com/assets/81969/11430906/be171070-945b-11e5-8d03-6a47800e4b72.png">

Testing:
- [x] `grunt watch` for SCSS files
- [x] `grunt watch` for JS files
- [x] `grunt watch` for Handlebars