@MariagraziaAlastra 2015-11-30
@jagtalon Did you perhaps commit and push changes from other branches you were on, here? It looks like you added some SCSS changes in here as well