@jagtalon 2015-11-30
@MariagraziaAlastra yeah I apologize for the mess I started doing the dev touchups in here as well since it was so nice to have something running quickly. I'll find a way to separate the two